+++ ModTek/Features/Logging/HarmonyXLoggerAdapter.cs
@@ -1,18 +1,18 @@
 ﻿using HBS.Logging;
-using NullableLogging;
+using ModTek.Public;
 using Logger = HarmonyLib.Tools.Logger;
 
 namespace ModTek.Features.Logging;
 
 internal static class HarmonyXLoggerAdapter
 {
-    private static readonly HBS.Logging.Logger.LogImpl s_logImpl = Log.HarmonyX.Log;
+    private static readonly ILog s_log = Log.HarmonyX.Log;
     internal static void Setup()
     {
         Logger.MessageReceived += (_, args) =>
         {
             var level = MapHarmonyLogChannelToHbsLogLevel(args.LogChannel);
-            s_logImpl.LogAtLevel(level, args.Message);
+            s_log.LogAtLevel(level, args.Message);
         };
     }