+++ ModTek/Log.cs
@@ -1,5 +1,5 @@
 ﻿using System.Diagnostics;
-using NullableLogging;
+using ModTek.Public;
 
 namespace ModTek;